requirements.txt
anyio==2.2.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
blis==0.7.4
catalogue==2.0.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cycler==0.10.0
cymem==2.0.5
decorator==5.0.7
defusedxml==0.7.1
deprecation==2.1.0
entrypoints==0.3
idna==2.10
ipykernel==5.5.4
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-http-over-ws==0.0.8
jupyter-packaging==0.10.1
jupyter-server==1.6.4
jupyterlab==3.0.14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.5.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
murmurhash==1.0.5
nbclassic==0.2.7
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.20.2
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pathy==0.5.2
photon==1.1.8.3
pickleshare==0.7.5
Pillow==8.2.0
pl-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/pl_core_news_lg-3.0.0/pl_core_news_lg-3.0.0-py3-none-any.whl
preprocessor==1.1.3
preshed==3.0.5
prometheus-client==0.10.1
prompt-toolkit==3.0.18
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.7.3
Pygments==2.9.0
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
pywinpty==1.0.1
pyzmq==22.0.3
requests==2.25.1
scikit-learn==0.24.2
scikit-plot==0.3.7
scipy==1.6.3
seaborn==0.11.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.16.0
sklearn==0.0
smart-open==3.0.0
sniffio==1.2.0
soupsieve==2.2.1
spacy==3.0.6
spacy-legacy==3.0.5
srsly==2.4.1
terminado==0.9.4
testpath==0.4.4
thinc==8.0.3
threadpoolctl==2.1.0
tld==0.12.5
toml==0.10.1
tomlkit==0.7.0
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
tweet-preprocessor==0.6.0
typer==0.3.2
unicode==2.8
Unidecode==1.2.0
urllib3==1.26.4
wasabi==0.8.2
wcwidth==0.2.5
webdriver-manager==3.4.1
webencodings==0.5.1
wordcloud==1.8.1
wrapt==1.12.1
